splits preprocess in preprocess and detect
